Rename allow-snapshot directory
* Renamed allow-snapshot directory
  to hdfs-allow-snapshot.
