+++ hadoop-hdfs-project/hadoop-hdfs-native-client/src/main/native/libhdfspp/tools/hdfs-allow-snapshot/hdfs-allow-snapshot.cc
@@ -25,7 +25,7 @@
 
 #include <google/protobuf/stubs/common.h>
 
-#include "allow-snapshot.h"
+#include "hdfs-allow-snapshot.h"
 #include "tools_common.h"
 
 HdfsAllowSnapshot::HdfsAllowSnapshot(const int argc, char **argv)
